fsinfo: Fix a couple of warnings
Fix a couple of printk format warnings where %llx is matched up with __u64,
but __u64 is unsigned long, not unsigned long long.

samples/vfs/test-mntinfo.c: In function 'display_mount':
samples/vfs/test-mntinfo.c:146:19: warning: format '%llx' expects argument of type 'long long unsigned int', but argument 3 has type '__u64' {aka 'long unsigned int'} [-Wformat=]
samples/vfs/test-fsinfo.c: In function 'dump_fsinfo_generic_mount_all':
samples/vfs/test-fsinfo.c:384:30: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 5 has type '__u64' {aka 'long unsigned int'} [-Wformat=]

Reported-by: Stephen Rothwell <sfr@canb.auug.org.au>

+++ samples/vfs/test-fsinfo.c
@@ -383,7 +383,7 @@ static void dump_fsinfo_generic_mount_all(void *reply, unsigned int size)
 
 	printf("%5x %5x %12llx %10llu %s\n",
 	       r->mnt_id, r->parent_id, (unsigned long long)r->mnt_unique_id,
-	       r->mnt_notify_sum, mp);
+	       (unsigned long long)r->mnt_notify_sum, mp);
 }
 
 static void dump_afs_fsinfo_server_address(void *reply, unsigned int size)
